Update and rename 5DChess.py to game.py
